+++ src/plone/app/multilingual/browser/helper_views.py
@@ -21,13 +21,20 @@ from zope.component import getMultiAdapter
 from zope.component import getUtility
 from zope.component.hooks import getSite
 from zope.interface import implementer
+from zope.interface import Interface
 from zope.publisher.interfaces import IPublishTraverse
 from zope.publisher.interfaces import NotFound
 
 try:
     from Products.ATContentTypes.interfaces.factory import IFactoryTool
 except ImportError:
-    from Products.CMFPlone.interfaces.factory import IFactoryTool
+    try:
+        from Products.CMFPlone.interfaces.factory import IFactoryTool
+    except ImportError:
+        # gone in Plone 5.1 w/o ATCT
+
+        class IFactoryTool(Interface):
+            pass
 
 
 @implementer(IPublishTraverse)